changed register/unregister interface in java to use longs instead of ints because pointers were getting truncated when passed from 64 bit jlongs to 32 bit jints. This wasn't getting noticed on windows because the address space was low enough, but it was showing up on other 64 bit systems.

+++ Java/Debugger/src/main/java/edu/umich/soar/debugger/modules/AbstractUpdateView.java
@@ -56,9 +56,9 @@ public abstract class AbstractUpdateView extends AbstractView  implements Agent.
 	
 	protected boolean   m_Updating = false ;
 	
-	protected int m_StopCallback = -1;
-	protected int m_InitCallback = -1;
-	protected int m_DecisionCallback = -1;
+	protected long m_StopCallback = -1;
+	protected long m_InitCallback = -1;
+	protected long m_DecisionCallback = -1;
 
 	// The constructor must take no arguments so it can be called
 	// from the loading code and the new window dialog
